third_party/WebKit/LayoutTests/fast/hidpi/image-srcset-simple-1x-expected.txt

PASS document.getElementById("foo").clientWidth==400 is true
PASS successfullyParsed is true

TEST COMPLETE
This test passes if the image below says 1x with a reddish background when the deviceScaleFactor is 1, and if says 2x with a greenish background when the deviceScaleFactor is 2.
